make install scripts idempotent
This allows a renamed library to be installed "over" an existing install.

Note that ALTER TABLE ... ADD CONSTRAINT does not support 'IF NOT EXISTS',
thus the constraint is dropped and re-added. This will revalidate the
constraint, but this should take a trivial amount of time as the metadata
tables are generally very small. Alternatively, NOT VALID could be added
to this command with a VALIDATE CONSTRAINT to be called later.